Port continuation (#3)
This PR addresses a bunch of Nits on the previous PR: #2 

Specifically:
- Instead of the integration test that tests a single sample, I added a
proptest.
- This :point_up_2: allowed me to remove the `Sampleable` stuff, since
that'd be duplicated between WNFS and this crate, which I don't think is
ideal. I'm not sure where to put the `Sampleable` trait best. Anyhow, we
don't really need it. Same with the `rvg` stuff.
- Changed any occurrences of `expede` in links/as the github owner to
`wnfs-wg`.
- Adjusted the docs in the -wasm crate to explain better what the const
params are.
- Some smaller nits such as how imports are made & some polishing.