Correctly handle empty bloom filters, avoid infinite loop (#9)
`HashIndexIterator` used to loop infinitely when `bit_size` was 0,
because it wouldn't be able to generate a random index that's `< 0`
(fair enough).

Now it simply exists instantly if `bit_size == 0`.

Also added a test case for empty blooms. They "technically" contain
everything as a false positive. I opted for that, rather than "an empty
bloom contains nothing", since that keeps the invariant that if you
`.insert` something into a bloom filter, it will *always* be
`contain`ed.